实现incr和decr，优化部分代码

+++ src/t_hash.go
@@ -44,5 +44,5 @@ func hDelCommand(c *SRedisClient) {
 			break
 		}
 	}
-	c.addReplyLongLong(deleted)
+	c.addReplyLongLong(int64(deleted))
 }
